Add Institute Page and template
Creates a base template "simple.html" to be used by other
templates. Create a Institute Page Model for the Institution
running courses.

Adds Department and Program models with snippets to create a
heirarchy for the courses.

Adds a logo to the homepage that can be shown in the navigation
menu.

Signed-off-by: Pradyumna Krishna <git@onpy.in>